pyproject.toml
[project]
name = "arknights-toolkit"
version = "0.4.0"
description = "Arknights-related Toolkit"
authors = [
    {name = "RF-Tar-Railt", email = "rf_tar_railt@qq.com"},
]
dependencies = [
    "Pillow>=9.2.0",
    "httpx>=0.23.0",
    "loguru>=0.6.0",
    "lxml>=4.9.1",
    "playwright>=1.26.1",
    "xlsxwriter>=3.0.3",
    "matplotlib>=3.6.2",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "AGPL-3.0"}

[project.urls]
repository = "https://github.com/RF-Tar-Railt/arknigths-toolkit"

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "black>=22.8.0",
    "isort>=5.10.1",
]

[tool.pdm.build]
includes = ["arknights_toolkit"]
excludes = [
    "arknights_toolkit/resource/ops_initialized",
    "arknights_toolkit/resource/careers.json",
    "arknights_toolkit/resource/operators/*"
]
